usecase/document-based-application/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
openai==1.52.2
langchain==0.3.23
python-dotenv==1.0.1
pydantic==2.6.4